Backend/requirements.txt
python-dotenv~=0.20.0
Flask~=2.1.2